Slide locations, new zoom control, css update
Fixed the slides sequence so that each slide represents a collection of administrative polygons and/or the archaeological site points. I've created a new zoom control within the bottom jQuery call for the 'cat_slides' data to update the zoom extent based on a proper scale that fits those administrative boundaries. Added some basic styling to the 'historyBox' div element but might remove that later.